app/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
black==24.2.0
certifi==2024.2.2
click==8.1.7
databases==0.9.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
orjson==3.9.15
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
ujson==5.9.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
